apps/streaming/serializers/category.py
from rest_framework import serializers
from apps.streaming.models import Category

class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = '__all__'