requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.10.0
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
azure-common==1.1.28
azure-communication-email==1.1.0
azure-core==1.36.0
azure-identity==1.25.1
azure-mgmt-communication==2.1.0
azure-mgmt-core==1.6.0
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
channels==4.3.1
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.3
daphne==3.0.2
Django==5.2.8
django-background-tasks==1.2.8
django-celery==3.1.17
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-rest-framework==0.1.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.1
fuzzywuzzy==0.18.0
gunicorn==23.0.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
isodate==0.7.2
kombu==5.5.4
lml==0.1.0
msal==1.34.0
msal-extensions==1.3.1
msgpack==1.1.0
# Editable Git install with no remote (ntc==1.0.0)
-e c:\users\user\documents\projects\ntc-cloud\scripts\ntc
openai==1.63.2
packaging==25.0
pandas==2.0.3
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.23
pyexcel==0.7.0
pyexcel-io==0.6.6
PyJWT==1.7.1
pyOpenSSL==25.0.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.2.1
requests==2.32.5
service-identity==24.2.0
setuptools==69.0.3
six==1.17.0
sqlparse==0.5.3
texttable==1.7.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
zope.interface==7.2
ffmpeg-python==0.2.0
